Constrain dependencies
Prior, was not using consraints for dependencies which causes
install to use the latest packages. This can cause inconsistent
builds at different times or on different machines despite
using the same config. Fluent and metadata filter are being
upgraded to match upstream and resolve pods restarting due to
not handling closed connections properly.